refactor: separate the file creation functionality and use it in app-sidebar and file tree components
