+++ src/renderer/src/components/create-file/index.ts
@@ -1,0 +1,1 @@
+export { CreateFileDialog } from "./components/create-file.dialog.component";
